+++ src/NavigationMenu/NavFooter.test.tsx
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
@@ -6,7 +6,7 @@ import { NavFooter } from './NavFooter.tsx';
 
 describe('NavFooter', () => {
   it('should display CollapseIcon and title "Collapse" when isNavExpanded is true', () => {
-    render(<NavFooter isNavExpanded toggleNavExpanded={() => {}} />);
+    render(<NavFooter isNavExpanded={true} toggleNavExpanded={() => {}} />);
 
     expect(screen.getByTitle('Collapse')).toBeVisible();
     expect(screen.getByTestId('collapse-icon')).toBeVisible();
@@ -23,23 +23,9 @@ describe('NavFooter', () => {
 
   it('should call toggleNavExpanded when button is clicked', () => {
     const toggleNavExpanded = vi.fn();
-    render(<NavFooter isNavExpanded toggleNavExpanded={toggleNavExpanded} />);
+    render(<NavFooter isNavExpanded={false} toggleNavExpanded={toggleNavExpanded} />);
     fireEvent.click(screen.getByRole('button'));
 
     expect(toggleNavExpanded).toHaveBeenCalled();
-  });
-
-  it('should display PlatformIcon when Expanded and text', () => {
-    render(<NavFooter isNavExpanded toggleNavExpanded={() => {}} />);
-
-    expect(screen.getByTestId('platform-icon')).toBeVisible();
-    expect(screen.queryByText('Platform')).toBeVisible();
-  });
-
-  it('should display PlatformIcon when collapsed and no text', () => {
-    render(<NavFooter isNavExpanded={false} toggleNavExpanded={() => {}} />);
-
-    expect(screen.getByTestId('platform-icon')).toBeVisible();
-    expect(screen.queryByText('Platform')).not.toBeInTheDocument();
   });
 });