+++ src/index.ts
@@ -23,6 +23,8 @@ import { Label, FieldSet } from './Form/index.ts';
 import { Select, SelectButton, SelectOption, SelectOptions, SelectOptionGroup } from './Select/index.ts';
 import { Alert } from './Alert/Alert.tsx';
 import { ActionableAlert } from './ActionableAlert/ActionableAlert.tsx';
+import { AccountMenu } from './AccountMenu/AccountMenu.tsx';
+import { Slider } from './Slider/Slider.tsx';
 
 export { Card, CardBody, CardFooter, CardHeader };
 export { Button };
@@ -46,6 +48,8 @@ export { Avatar };
 export { Menu, MenuButton, MenuItem, MenuItems };
 export { Alert, ActionableAlert };
 export { NavigationMenu };
+export { AccountMenu };
+export { Slider };
 
 // Export the tailwind plugin
 export default tailwindPlugin;